Split --verify into --verify-decode & --verify-duration
Drop the matches_input heuristic that silently skipped checks for
customised encodes. Each check is now explicit and deterministic,
--verify enables both. The audio stream check is removed, there was
no evidence of ffmpeg ever dropping it.

Move the decode pass into the ffmpeg module, show verify as the last
third of the encode progress bar and document the 2s duration tolerance.
