+++ README.md
@@ -59,9 +59,11 @@ ab-av1 encode [OPTIONS] -i <INPUT> --crf <CRF> --preset <PRESET>
 
 #### Notable options
 These also apply to [auto-encode](#command-auto-encode).
-* `--verify` decodes the finished encode & fails on decode errors, a duration mismatch with
-  the input or missing audio. The comparisons against the input are skipped for customised
-  encodes, e.g. when using `--vfilter` or `--enc`. Costs an extra full decode.
+* `--verify` checks the finished encode before it is moved into place: decodes it in full
+  failing on decode errors (`--verify-decode`) & fails if its duration does not match the
+  input within 2s (`--verify-duration`). Either check can be enabled on its own.
+  The encoder can otherwise exit successfully having written a damaged or cut short result,
+  e.g. from a truncated input.
 * `--fail-fast` stops the encode at the first ffmpeg reported error instead of writing a
   damaged result.
