+++ src/main.rs
@@ -7,7 +7,6 @@ mod log;
 mod process;
 mod sample;
 mod temporary;
-mod verify;
 mod vmaf;
 mod xpsnr;
